python/requirements.txt
web3~=5.28.0
hexbytes~=0.2.2
numpy~=1.22.3
dacite~=1.6.0
aiohttp~=3.8.1
requests~=2.27.1
python-dateutil~=2.8.2
cachetools~=5.0.0
discord-webhook[async]~=0.17.0
absl-py~=1.2.0
ratelimit~=2.2.1
py-solc-x~=1.1.1
python-dotenv~=0.21.0